requirements.txt
openai
anthropic
groq
python-dotenv
langchain
langchain-openai
llama-index